Move job list so the list isn't fetched until after auth
